# swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the FrescosAPI documentation.",
        "title": "FrescosAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/sellers": {
            "get": {
                "description": "This endpoint retrieves all sellers from the database using the service layer.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sellers"
                ],
                "summary": "Retrieve all sellers",
                "responses": {
                    "200": {
                        "description": "OK - The sellers were successfully retrieved",
                        "schema": {
                            "$ref": "#/definitions/sellersctl.SellerResJSON"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - An unexpected error occurred during the retrieval process",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "This endpoint creates a new seller based on the provided JSON request body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sellers"
                ],
                "summary": "Create a new seller",
                "parameters": [
                    {
                        "description": "Seller Create JSON",
                        "name": "seller",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sellersctl.SellerCreateJSON"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Seller created successfully",
                        "schema": {
                            "$ref": "#/definitions/sellersctl.SellerResJSON"
                        }
                    },
                    "400": {
                        "description": "Invalid request data or JSON decoding error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict - Duplicate entry",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/sellers/{id}": {
            "get": {
                "description": "This endpoint retrieves a seller based on the provided ID in the URL parameter.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sellers"
                ],
                "summary": "Get a seller by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Seller ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - The seller was successfully retrieved",
                        "schema": {
                            "$ref": "#/definitions/sellersctl.SellerResJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request - The request ID is invalid or less than 1",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - The seller with the specified ID does not exist",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "This endpoint deletes a seller based on the provided ID in the URL parameter.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sellers"
                ],
                "summary": "Delete a seller",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Seller ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content - The seller was successfully deleted"
                    },
                    "400": {
                        "description": "Bad Request - The request ID is invalid or less than 1",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - The seller with the specified ID does not exist",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict - The seller cannot be deleted due to a MySQL foreign key constraint error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - An unexpected error occurred during the deletion process",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "This endpoint updates a seller based on the provided ID in the URL parameter and the JSON request body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sellers"
                ],
                "summary": "Update a seller",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Seller ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Seller update data",
                        "name": "seller",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sellersctl.SellerUpdateJSON"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - The seller was successfully updated",
                        "schema": {
                            "$ref": "#/definitions/sellersctl.SellerResJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request - The seller cannot be updated due to a MySQL cannot add or update child row error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict - The seller cannot be updated due to a MySQL duplicate entry error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - An unexpected error occurred during the update process",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "sellersctl.SellerCreateJSON": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "cid": {
                    "type": "string"
                },
                "company_name": {
                    "type": "string"
                },
                "locality_id": {
                    "type": "integer"
                },
                "telephone": {
                    "type": "string"
                }
            }
        },
        "sellersctl.SellerResJSON": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "sellersctl.SellerUpdateJSON": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "cid": {
                    "type": "string"
                },
                "company_name": {
                    "type": "string"
                },
                "locality_id": {
                    "type": "integer"
                },
                "telephone": {
                    "type": "string"
                }
            }
        },
...
    }
}